Moved files to src folder
